Remove warning from BLACKLIST definiations in alloc.c

+++ Server/src/alloc.c
@@ -473,8 +473,8 @@ showBlacklistStats(dbref player)
    double i_diver;
    char *s_buff, c_chr;
    
-   i_blsize = sizeof(BLACKLIST) * mudstate.blacklist_cnt;
-   i_ndsize = sizeof(BLACKLIST) * mudstate.blacklist_nodns_cnt;
+   i_blsize = (int)sizeof(BLACKLIST) * mudstate.blacklist_cnt;
+   i_ndsize = (int)sizeof(BLACKLIST) * mudstate.blacklist_nodns_cnt;
    s_buff = alloc_mbuf("blacklist_stats");
    
    notify(player, "\nBlacklist Stats    Size   Inuse     Total Mem (Bytes)");
@@ -492,7 +492,7 @@ showBlacklistStats(dbref player)
       i_diver = (double) i_blsize;
       c_chr = ' ';
    }
-   sprintf(s_buff, "%-18s %-6d %-9d %-12d (%.2f%c)", (char *)"Black List", sizeof(BLACKLIST), mudstate.blacklist_cnt, i_blsize, i_diver, c_chr);
+   sprintf(s_buff, "%-18s %-6d %-9d %-12d (%.2f%c)", (char *)"Black List", (int)sizeof(BLACKLIST), mudstate.blacklist_cnt, i_blsize, i_diver, c_chr);
    notify(player, s_buff);
    if ( i_ndsize > 1000000000 ) {
       i_diver = (double) i_ndsize / 1000000000.0;
@@ -507,7 +507,7 @@ showBlacklistStats(dbref player)
       i_diver = (double) i_ndsize;
       c_chr = ' ';
    }
-   sprintf(s_buff, "%-18s %-6d %-9d %-12d (%.2f%c)", (char *)"NoDNS list", sizeof(BLACKLIST), mudstate.blacklist_nodns_cnt, i_ndsize, i_diver, c_chr);
+   sprintf(s_buff, "%-18s %-6d %-9d %-12d (%.2f%c)", (char *)"NoDNS list", (int)sizeof(BLACKLIST), mudstate.blacklist_nodns_cnt, i_ndsize, i_diver, c_chr);
    notify(player, s_buff);
    free_mbuf(s_buff); 
 }
